refactor: slice A topic-split e2e support fat lib modules (#4195)
Split the four >400-line god helpers (naval move, fleet-reach NW
predicates, bootstrap, hit-testable) into concern-scoped siblings with
thin aggregators; preserve public entrypoints via barrels.

Refs #4195
