+++ packages/colonizethis_app_e2e_support/lib/e2e_test_shared.dart
@@ -25,6 +25,7 @@ export 'e2e_test_shared_panel_open_post_tap_probe.dart';
 export 'e2e_test_shared_panel_open_sheet_close.dart';
 export 'e2e_test_shared_panel_open_trigger_attempt.dart';
 export 'e2e_test_shared_panel_text_match.dart';
+export 'e2e_test_shared_naval_move.dart';
 export 'e2e_test_shared_panels.dart';
 export 'e2e_test_shared_region_tabs.dart';
 export 'e2e_test_shared_standard_scenario_opener.dart';
