+++ packages/colonizethis_app_e2e_support/lib/e2e_test_shared_bootstrap.dart
@@ -1,452 +1,6 @@
-import 'dart:async';
-import 'dart:ui' as ui;
-
-import 'package:colonizethis_app/features/game/flame/caches/civilian_icon_cache.dart';
-import 'package:colonizethis_app/features/game/flame/caches/fleet_icon_cache.dart';
-import 'package:colonizethis_app/features/game/screens/game/game_screen_shared.dart';
-import 'package:colonizethis_app/features/game/flame/caches/province_label_icon_cache.dart';
-import 'package:colonizethis_app/features/game/flame/caches/resource_icon_cache.dart';
-import 'package:colonizethis_app/features/game/flame/caches/town_icon_cache.dart';
-import 'package:colonizethis_app/widgets/ct_dialog_shell.dart';
-import 'package:colonizethis_app/widgets/ct_dropdown.dart';
-import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:colonizethis_app_e2e_support/e2e_support_validation_exception.dart';
-import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
-
-/// Loads and decodes each path with bounded concurrency (overlapping I/O +
-/// image decode completion) instead of strictly serial awaits.
-///
-/// Larger batches reduce the number of synchronization points where the loop
-/// awaits *all* in-flight decodes before scheduling the next group, so a slow
-/// decode no longer blocks every other decode in the same batch. The default
-/// is sized to comfortably cover the relocated 64px icon manifest (~60 assets)
-/// in a single batch, while still allowing callers on memory-constrained CI
-/// runners to opt back into smaller chunks. Refs GitHub #2336 AC3.
-Future<List<String>> e2eDecodePngAssetPathsParallel(
-  List<String> assetPaths, {
-  int batchSize = 64,
-}) async {
-  if (batchSize <= 0) {
-    throw E2eSupportValidationException('batchSize must be positive');
-  }
-  final failures = <String>[];
-  for (var i = 0; i < assetPaths.length; i += batchSize) {
-    final end = i + batchSize > assetPaths.length
-        ? assetPaths.length
-        : i + batchSize;
-    final chunk = assetPaths.sublist(i, end);
-    final chunkFailures = await Future.wait(
-      chunk.map((assetPath) async {
-        try {
-          final data = await rootBundle.load(assetPath);
-          final bytes = data.buffer.asUint8List();
-          final completer = Completer<ui.Image>();
-          ui.decodeImageFromList(bytes, completer.complete);
-          final image = await completer.future;
-          image.dispose();
-          return null;
-        } catch (e) {
-          return '$assetPath ($e)';
-        }
-      }),
-    );
-    for (final message in chunkFailures) {
-      if (message != null) {
-        failures.add(message);
-      }
-    }
-  }
-  return failures;
-}
-
-/// Default phase label emitted by [e2eWaitForMapHudAfterNewGameStart] when a
-/// caller does not override [E2ePerfLog] attribution.
-///
-/// The constant is exposed so widget-test pins and downstream perf-marker
-/// scrapers (for example the GitHub #2336 AC8 baseline timing tool) can refer
-/// to the legacy label by name instead of hard-coding the literal string.
-const String kE2eDefaultWaitForMapHudPhase =
-    'wait_for_map_hud_after_new_game_start';
-
-/// Counter name bumped on every [e2eWaitForMapHudAfterNewGameStart] iteration
-/// so a hung bootstrap surfaces as an attributable counter spike instead of a
-/// silent 60 s wall-clock burn (Refs GitHub #2336 AC8 / AC10).
-const String kE2eWaitForMapHudIterationsCounter =
-    'wait_for_map_hud_after_new_game_start_iterations';
-
-/// After [Start] is tapped, polls until the in-game map HUD is visible.
-///
-/// Evaluates success before the first pump; uses [e2eNextIdlePollStepMs]
-/// (25ms → doubling capped at 500ms) per
-/// `SPEC/program/e2e-integration-tests.md` § *Adaptive poll pacing* /
-/// GitHub #2336 AC5 adaptive polling guidance.
-///
-/// Perf attribution (Refs GitHub #2336 AC8 / baseline measurement):
-///
-/// - When [perf] is non-null, emits one `E2E_TIMING|phase=[phaseName]` line
-///   on every return path with a `result=...` meta tag distinguishing
-///   `result=already_mounted` (entry-iteration short-circuit; counter
-///   value `1`), `result=advanced` (HUD landed on a later iteration after
-///   one or more pumps), `result=error_dialog` (fail-fast on
-///   `Could not create game`), and `result=timeout` (overall-cap fail path).
-///   The phase label defaults to [kE2eDefaultWaitForMapHudPhase].
-/// - Bumps the [kE2eWaitForMapHudIterationsCounter] counter once per loop
-///   iteration (including the iteration that returns success) so a hung
-///   bootstrap surfaces as a counter spike instead of a silent wall-clock
-///   burn. The counter is incremented before the early-exit checks for that
-///   iteration so the `result=already_mounted` short-circuit reports
-///   `value=1`.
-Future<void> e2eWaitForMapHudAfterNewGameStart(
-  WidgetTester tester, {
-  Duration overallCap = const Duration(seconds: 60),
-  E2ePerfLog? perf,
-  String phaseName = kE2eDefaultWaitForMapHudPhase,
-}) async {
-  final sw = Stopwatch()..start();
-  final setupDeadline = DateTime.now().add(overallCap);
-  var stepMs = 25;
-  var iterations = 0;
-  while (DateTime.now().isBefore(setupDeadline)) {
-    iterations += 1;
-    perf?.bumpCounter(
-      kE2eWaitForMapHudIterationsCounter,
-      meta: 'phase=$phaseName',
-    );
-    if (find.text('Could not create game').evaluate().isNotEmpty) {
-      perf?.timing(phaseName, sw.elapsed, meta: 'result=error_dialog');
-      fail(
-        'New game setup failed (error dialog). '
-        'Exception: ${tester.takeException()}',
-      );
-    }
-    if (find.byKey(kHomeToCapitalButtonKey).evaluate().isNotEmpty) {
-      await e2eAdvanceGameStartIntroUntilDismissed(tester, perf: perf);
-      perf?.timing(
-        phaseName,
-        sw.elapsed,
-        meta: iterations == 1 ? 'result=already_mounted' : 'result=advanced',
-      );
-      return;
-    }
-    if (e2eGameStartIntroBlocksUi(tester)) {
-      await e2eAdvanceGameStartIntroUntilDismissed(tester, perf: perf);
-      stepMs = 25;
-      continue;
-    }
-    if (find.text('Creating game').evaluate().isNotEmpty) {
-      await tester.pump(Duration(milliseconds: stepMs));
-      stepMs = e2eNextIdlePollStepMs(stepMs);
-      continue;
-    }
-    await tester.pump(Duration(milliseconds: stepMs));
-    stepMs = e2eNextIdlePollStepMs(stepMs);
-  }
-  perf?.timing(phaseName, sw.elapsed, meta: 'result=timeout');
-  fail(
-    'Timed out after ${overallCap.inSeconds}s waiting for '
-    'map (home→capital). Last exception: ${tester.takeException()}',
-  );
-}
-
-/// Selects an advanced-start preset in DLG10001 before tapping **Start**.
-///
-/// Requires the locked full-init profile (`CT_E2E_LOCKED_FULL_INIT=true`).
-/// [optionLabel] must match the English l10n string (e.g. `50 Turns In (1598)`).
-Future<void> e2eSelectLeaderDialogAdvancedStart(
-  WidgetTester tester,
-  String optionLabel, {
-  E2ePerfLog? perf,
-}) async {
-  final dropdown = find.byType(CtDropdown<AdvancedStartType>);
-  expect(dropdown, findsOneWidget);
-
-  // Locked full-init DLG10001 exceeds the 720 px CI viewport; scroll the shell
-  // body before tapping (same scrollable as the Start-button drag below).
-  final shellScrollable = find.descendant(
-    of: find.byType(CtDialogShell),
-    matching: find.byType(Scrollable),
-  );
-  await tester.dragUntilVisible(
-    dropdown,
-    shellScrollable,
-    const Offset(0, -120),
-  );
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () => dropdown.hitTestable().evaluate().isNotEmpty,
-    timeout: const Duration(milliseconds: 600),
-    perf: perf,
-    phaseName: 'pump_until_advanced_start_dropdown_tappable',
-  );
-
-  final didOpenMenu = await e2eEnsureVisibleAndTapHitTestable(
-    tester,
-    dropdown,
-  );
-  expect(
-    didOpenMenu,
-    isTrue,
-    reason: 'Advanced start dropdown must be tappable after scroll',
-  );
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () => find.text(optionLabel).evaluate().isNotEmpty,
-    timeout: const Duration(seconds: 5),
-    perf: perf,
-    phaseName: 'pump_until_advanced_start_menu_open',
-  );
-
-  final menuOption = find.text(optionLabel);
-  expect(
-    menuOption,
-    findsWidgets,
-    reason: 'Advanced start menu must list $optionLabel',
-  );
-  final didSelectOption = await e2eEnsureVisibleAndTapHitTestable(
-    tester,
-    menuOption.last,
-  );
-  expect(
-    didSelectOption,
-    isTrue,
-    reason: 'Advanced start option $optionLabel must be tappable',
-  );
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () => find
-        .widgetWithText(CtDropdown<AdvancedStartType>, optionLabel)
-        .evaluate()
-        .isNotEmpty,
-    timeout: const Duration(seconds: 5),
-    perf: perf,
-    phaseName: 'pump_until_advanced_start_selected',
-  );
-}
-
-/// Canonical new-game → map HUD path shared by E2E scenarios (Refs #2336).
-///
-/// Adaptive polling notes (GitHub #2336 AC5):
-///
-/// - After the `New Game` tap, [e2eWaitUntilFound] polls for the `Start`
-///   button with check-before-pump + exponential backoff (25 → 500 ms cap)
-///   via the inner [e2eWaitUntilFound] call.
-/// - After the drag that scrolls the dialog to the `Start` button,
-///   [e2ePumpUntilConditionOrIdle] polls for the button becoming
-///   hit-testable with the same check-before-pump cadence.
-/// - After the `Start` tap, [e2ePumpUntilConditionOrIdle] polls for any
-///   post-tap observable (`Creating game` indicator, the home-to-capital
-///   button on instant map gen, the intro-overlay loading branch, or the
-///   `Could not create game` error dialog) with a short 200 ms safety net.
-///   The downstream [e2eWaitForMapHudAfterNewGameStart] then handles the
-///   longer-running adaptive poll up to [overallCap]. The settle slice is
-///   attributable to `pump_until_post_start_tap_settled`.
-/// - After the map HUD mounts, [e2ePumpUntilConditionOrIdle] polls for the
-///   home-to-capital button becoming hit-testable before
-///   [e2eAdvanceGameStartIntroUntilDismissed] dismisses the intro overlay.
-///
-/// No bare single-frame `tester.pump()` settles remain in the body; every
-/// state transition is gated by a condition-based wait so the helper
-/// conforms to AC5 end-to-end.
-Future<void> e2eBootstrapNewGameToMap(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration overallCap = const Duration(seconds: 60),
-  String? advancedStartOptionLabel,
-}) async {
-  final phaseSw = Stopwatch()..start();
-  await tester.tap(find.text('New Game'));
-  await e2eWaitUntilFound(
-    tester,
-    find.text('Start'),
-    timeout: const Duration(seconds: 30),
-    perf: perf,
-    phaseName: 'wait_until_found_start_button',
-  );
-
-  if (advancedStartOptionLabel != null) {
-    await e2eSelectLeaderDialogAdvancedStart(
-      tester,
-      advancedStartOptionLabel,
-      perf: perf,
-    );
-  }
-
-  final startButton = find.ancestor(
-    of: find.text('Start'),
-    matching: find.byType(CtNinePatchButton),
-  );
-  expect(startButton, findsOneWidget);
-
-  final shellScrollable = find.descendant(
-    of: find.byType(CtDialogShell),
-    matching: find.byType(Scrollable),
-  );
-  await tester.dragUntilVisible(
-    startButton,
-    shellScrollable,
-    const Offset(0, -120),
-  );
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () => startButton.hitTestable().evaluate().isNotEmpty,
-    timeout: const Duration(milliseconds: 600),
-    perf: perf,
-    phaseName: 'pump_until_start_button_tappable_after_drag',
-  );
-  await tester.ensureVisible(startButton);
-  await tester.tap(startButton);
-  // Replace the legacy bare single-frame `await tester.pump();` settle with
-  // an adaptive condition-based wait so the helper conforms to GitHub #2336
-  // AC5 (check-before-pump + exponential backoff capped at ≤500 ms). The
-  // poll short-circuits the moment any post-Start-tap observable surfaces
-  // (`Creating game` indicator, the home-to-capital button on instant map
-  // gen, the intro-overlay loading branch, or the `Could not create game`
-  // error dialog), so the canonical success path (which mounts `Creating
-  // game` on the very next frame) returns within a single 25 ms pump
-  // instead of opaque pump-then-blind-handoff sequencing. The downstream
-  // [e2eWaitForMapHudAfterNewGameStart] still runs the longer-running
-  // adaptive poll; this settle is a short, observable safety net so
-  // post-Start-tap latency is attributable to a dedicated
-  // `pump_until_post_start_tap_settled` perf slice. The 200 ms cap is
-  // strictly a fast-path bound — the downstream helper's `overallCap`
-  // (default 60 s) governs the slow path.
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () =>
-        find.text('Creating game').evaluate().isNotEmpty ||
-        find.byKey(kHomeToCapitalButtonKey).evaluate().isNotEmpty ||
-        find.text('Could not create game').evaluate().isNotEmpty ||
-        e2eGameStartIntroBlocksUi(tester),
-    timeout: const Duration(milliseconds: 200),
-    perf: perf,
-    phaseName: 'pump_until_post_start_tap_settled',
-  );
-
-  await e2eWaitForMapHudAfterNewGameStart(
-    tester,
-    overallCap: overallCap,
-    perf: perf,
-  );
-
-  expect(find.byKey(kHomeToCapitalButtonKey), findsOneWidget);
-  await e2ePumpUntilConditionOrIdle(
-    tester,
-    () =>
-        find.byKey(kHomeToCapitalButtonKey).hitTestable().evaluate().isNotEmpty,
-    timeout: const Duration(milliseconds: 800),
-    perf: perf,
-    phaseName: 'pump_until_home_capital_tappable_after_map',
-  );
-  await e2eAdvanceGameStartIntroUntilDismissed(tester, perf: perf);
-  perf?.timing('new_game_to_map', phaseSw.elapsed);
-}
-
-/// Collects non-empty [Text] data in depth-first preorder (E2E snapshot helpers).
-void e2eCollectTextPreorder(Element element, List<String> out) {
-  final w = element.widget;
-  if (w is Text) {
-    final d = w.data;
-    if (d != null && d.isNotEmpty) {
-      out.add(d);
-    }
-  }
-  element.visitChildren((child) {
-    e2eCollectTextPreorder(child, out);
-  });
-}
-
-/// Relocated 64px map icon paths from the asset manifest (sorted).
-Future<List<String>> e2eDiscoverRelocated64pxPngAssets() async {
-  final manifest = await AssetManifest.loadFromAssetBundle(rootBundle);
-  final assets =
-      manifest
-          .listAssets()
-          .where(
-            (assetPath) =>
-                assetPath.startsWith('assets/icons/64/') &&
-                assetPath.endsWith('.png'),
-          )
-          .toList()
-        ..sort();
-  return assets;
-}
-
-/// Asserts manifest contents match [expectedAssets], then decodes via
-/// [e2eDecodePngAssetPathsParallel] (Refs #2336 AC3).
-Future<void> e2eEnsureRelocated64pxPngDecode(
-  Set<String> expectedAssets, {
-  String emptyManifestReason =
-      'Expected relocated map icon PNGs under assets/icons/64/, but none were found in the asset manifest.',
-  String? countMismatchReason,
-  String? orderedMismatchReason,
-  String decodeFailuresPrefix =
-      'Failed to load one or more relocated 64px PNG assets:',
-}) async {
-  final assets = await e2eDiscoverRelocated64pxPngAssets();
-  final expectedSorted = expectedAssets.toList()..sort();
-  expect(assets, isNotEmpty, reason: emptyManifestReason);
-  expect(
-    assets.length,
-    expectedAssets.length,
-    reason:
-        countMismatchReason ??
-        'Unexpected number of relocated 64px PNG assets. '
-            'Expected ${expectedAssets.length} map-family files, found ${assets.length}.',
-  );
-  expect(
-    assets,
-    orderedEquals(expectedSorted),
-    reason:
-        orderedMismatchReason ??
-        'Relocated 64px PNG manifest entries do not match expected map icon families.',
-  );
-  final failures = await e2eDecodePngAssetPathsParallel(assets);
-  expect(
-    failures,
-    isEmpty,
-    reason: failures.isEmpty
-        ? null
-        : '$decodeFailuresPrefix\n${failures.join('\n')}',
-  );
-}
-
-/// Asserts the manifest matches the canonical map icon families, then decodes
-/// every PNG via [e2eDecodePngAssetPathsParallel] (bounded concurrency).
-///
-/// Used by new-game E2E tests that need the same warm-cache behavior.
-///
-/// Prefer [e2eEnsureAllRelocated64pxPngsLoadSuiteOnce] when each scenario needs
-/// the same decode/assert path so work runs at most once per test VM (GitHub
-/// #2336 AC3).
-Future<void> e2eEnsureAllRelocated64pxPngsLoad() async {
-  await e2eEnsureRelocated64pxPngDecode(<String>{
-    ...kCivilianIconSlugs.map(
-      (slug) => 'assets/icons/64/ui_icon_civ_$slug.png',
-    ),
-    ...kResourceIconIds.map(
-      (resourceId) => 'assets/icons/64/ui_icon_com_$resourceId.png',
-    ),
-    ...kTownIconIds.map(townIconCache.assetPath),
-    ...TownIconCache.legacyTownIconAssetPaths,
-    ...kProvinceLabelIconIds.map(
-      (iconId) => 'assets/icons/64/ui_icon_$iconId.png',
-    ),
-    kFleetMapIcon64PngAssetPath,
-  });
-}
-
-Future<void>? _e2eAllRelocated64pxPngLoadSuiteFuture;
-
-/// Runs [e2eEnsureAllRelocated64pxPngsLoad] at most once per isolate.
-///
-/// Subsequent calls await the same future so multiple E2E scenarios in one run
-/// do not repeat manifest + parallel decode work (GitHub #2336 AC3).
-Future<void> e2eEnsureAllRelocated64pxPngsLoadSuiteOnce() async {
-  _e2eAllRelocated64pxPngLoadSuiteFuture ??= e2eEnsureAllRelocated64pxPngsLoad();
-  return _e2eAllRelocated64pxPngLoadSuiteFuture!;
-}
+// New-game bootstrap and PNG warm-cache helpers — topic-split aggregator
+// (Refs #4195 slice A).
+export 'e2e_test_shared_bootstrap_map_hud.dart';
+export 'e2e_test_shared_bootstrap_new_game.dart';
+export 'e2e_test_shared_bootstrap_png_assets.dart';
+export 'e2e_test_shared_bootstrap_png_decode.dart';
